+++ src/app/components/PersonalAvatar/types.ts
@@ -1,0 +1,4 @@
+import { AvatarSize } from "../Avatar/types"
+export interface PersonalAvatarProps {
+    size?:AvatarSize
+}