+++ src/app/context/PrincipalContext/hooks/useAlert/types.ts
@@ -1,0 +1,2 @@
+import useAlert from "./useAlert";
+export type ReturnTypeUseAlert = ReturnType<typeof useAlert>;