+++ src/app/context/PrincipalContext/hooks/useTheme/types.ts
@@ -1,0 +1,10 @@
+export type Theme = 'light' | 'dark';
+
+/**
+ * Valor expuesto por `ThemeContext`.
+ */
+export interface UseThemeReturn{
+  theme: Theme;
+  toggleTheme: () => void;
+  setDarkTheme: () => void
+}