+++ src/app/context/PrincipalContext/types.ts
@@ -1,0 +1,6 @@
+import { UseThemeReturn } from "./hooks/useTheme/types"
+import { ReturnTypeUseAlert } from "./hooks/useAlert/types"
+export interface PrincipalContextValue {
+  usePrincipalTheme: UseThemeReturn,
+  usePrincipalAlert:ReturnTypeUseAlert
+}