+++ src/app/layout.tsx
@@ -3,11 +3,12 @@
 import { Montserrat } from 'next/font/google'
 import localFont from 'next/font/local'
 import './globals.css'
-import { ThemeProvider } from './context/ThemeContext/ThemeContext'
-import ThemeInitializer from './context/ThemeContext/ThemeInitializet'
+import { PrincipalProvider } from './context/PrincipalContext/PrincipalContext'
+import ThemeInitializer from './context/PrincipalContext/utilities/ThemeInitializer'
 import ServiceWorkerRegister from './components/ServiceWorkerRegister/ServiceWorkerRegister'
 import type { Metadata } from 'next'
 import { AuthProvider } from './context/AuthContext/AuthContext'
+import { FirebaseProvider } from './context/FirebaseContext/FirebaseContext'
 const montserrat = Montserrat({
   subsets: ['latin'],
   weight: ['400', '500', '600'],
@@ -38,15 +39,18 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="es">
       <body className={`${montserrat.variable} ${nulshock.variable}`}>
-        <ThemeProvider>
+        <PrincipalProvider>
           <AuthProvider>
-            <ThemeInitializer />
-            <ServiceWorkerRegister />
+            <FirebaseProvider>
+              <ThemeInitializer />
+              <ServiceWorkerRegister />
 
-            {/* Sincroniza el atributo data-theme en <html> */}
-            {children}
+              {/* Sincroniza el atributo data-theme en <html> */}
+              {children}
+            </FirebaseProvider>
+
           </AuthProvider>
-        </ThemeProvider>
+        </PrincipalProvider>
       </body>
     </html>
   )