+++ src/app/login/types.ts
@@ -1,4 +1,0 @@
-import useLogin from "./hooks/useLogin";
-export type LoginPageProps = {
-  routerOverride?: ReturnType<typeof useLogin> extends (arg: infer R) => any ? R : never;
-};