+++ src/app/main-page/dashboard/page.tsx
@@ -1,9 +1,0 @@
-'use client';
-import FormTestPage from "@/app/components/DynamicForm/FormTestPage";
-export default function DashboardPage() {
-  return (
-    <div>
-      <FormTestPage/>
-    </div>
-  );
-}